+++ RozetkaReactTs/package.json
@@ -12,8 +12,10 @@
   "dependencies": {
     "@emotion/react": "^11.13.5",
     "@emotion/styled": "^11.13.5",
+    "@mui/icons-material": "^6.1.8",
     "@mui/joy": "^5.0.0-beta.48",
     "@mui/material": "^6.1.8",
+    "antd": "^5.22.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router-dom": "^7.0.1"
