+++ RozetkaReactTs/src/App.tsx
@@ -14,6 +14,8 @@
 
 import {Route, Routes} from "react-router-dom";
 import Layout from "./components/Layout.tsx";
+import Home from "./components/Home.tsx";
+import AboutUs from "./components/AboutUs.tsx";
 
 
 
@@ -21,6 +23,9 @@ export default function App() {
     return (
         <Routes>
             <Route path="/" element={<Layout/>}>
+                <Route index element={<Home />} />
+                <Route path="test" element={<Home />} />
+                <Route path="aboutus" element={<AboutUs />} />
             </Route>
         </Routes>
     );
