+++ RozetkaReactTs/src/components/AboutUs.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+
+
+
+const AboutUs: React.FC = () => {
+
+
+    return (
+        <h1>AboutUs</h1>
+    );
+};
+
+export default AboutUs;
